+++ python/docs_viewer.py
@@ -16,6 +16,7 @@ from urllib.parse import unquote
 import gi
 
 gi.require_version("Gtk", "3.0")
+gi.require_version("Gdk", "3.0")
 from gi.repository import Gdk, GLib, Gtk
 
 try: